pyproject.toml
[project]
name = "pdfimagerag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-chroma>=0.2.1",
    "langchain>=0.3.15",
    "langgraph>=0.2.68",
    "pillow>=11.1.0",
    "pymupdf>=1.25.2",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.3.2",
]
